+++ modules/pricefeed/utils_get_denom.go
@@ -1,0 +1,12 @@
+package pricefeed
+
+func GetDenom() (denom string) {
+	for _, token := range PricefeedCfg.Tokens {
+		for _, unit := range token.Units {
+			if unit.Exponent == 0 {
+				denom = unit.Denom
+			}
+		}
+	}
+	return denom
+}